% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggecdf-package.R
\docType{package}
\name{ggecdf-package}
\alias{ggecdf}
\alias{ggecdf-package}
\title{ggecdf: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
